Refactoring: separating logic from ui
